+++ socialaggregator/migrations/0016_auto__chg_field_ressource_image__add_index_ressource_image__chg_field_.py
@@ -1,0 +1,185 @@
+# -*- coding: utf-8 -*-
+from south.utils import datetime_utils as datetime
+from south.db import db
+from south.v2 import SchemaMigration
+from django.db import models
+
+
+class Migration(SchemaMigration):
+
+    def forwards(self, orm):
+
+        # Renaming column for 'Ressource.image' to match new field type.
+        db.rename_column(u'socialaggregator_ressource', 'image', 'image_id')
+        # Changing field 'Ressource.image'
+        db.alter_column(u'socialaggregator_ressource', 'image_id', self.gf('django.db.models.fields.related.ForeignKey')(null=True, to=orm['filer.Image']))
+        # Adding index on 'Ressource', fields ['image']
+        db.create_index(u'socialaggregator_ressource', ['image_id'])
+
+
+        # Renaming column for 'Ressource.thumbnail' to match new field type.
+        db.rename_column(u'socialaggregator_ressource', 'thumbnail', 'thumbnail_id')
+        # Changing field 'Ressource.thumbnail'
+        db.alter_column(u'socialaggregator_ressource', 'thumbnail_id', self.gf('django.db.models.fields.related.ForeignKey')(null=True, to=orm['filer.Image']))
+        # Adding index on 'Ressource', fields ['thumbnail']
+        db.create_index(u'socialaggregator_ressource', ['thumbnail_id'])
+
+
+    def backwards(self, orm):
+        # Removing index on 'Ressource', fields ['thumbnail']
+        db.delete_index(u'socialaggregator_ressource', ['thumbnail_id'])
+
+        # Removing index on 'Ressource', fields ['image']
+        db.delete_index(u'socialaggregator_ressource', ['image_id'])
+
+
+        # User chose to not deal with backwards NULL issues for 'Ressource.image'
+        raise RuntimeError("Cannot reverse this migration. 'Ressource.image' and its values cannot be restored.")
+        
+        # The following code is provided here to aid in writing a correct migration
+        # Renaming column for 'Ressource.image' to match new field type.
+        db.rename_column(u'socialaggregator_ressource', 'image_id', 'image')
+        # Changing field 'Ressource.image'
+        db.alter_column(u'socialaggregator_ressource', 'image', self.gf('django.db.models.fields.files.ImageField')(max_length=100))
+
+        # User chose to not deal with backwards NULL issues for 'Ressource.thumbnail'
+        raise RuntimeError("Cannot reverse this migration. 'Ressource.thumbnail' and its values cannot be restored.")
+        
+        # The following code is provided here to aid in writing a correct migration
+        # Renaming column for 'Ressource.thumbnail' to match new field type.
+        db.rename_column(u'socialaggregator_ressource', 'thumbnail_id', 'thumbnail')
+        # Changing field 'Ressource.thumbnail'
+        db.alter_column(u'socialaggregator_ressource', 'thumbnail', self.gf('django.db.models.fields.files.ImageField')(max_length=100))
+
+    models = {
+        u'auth.group': {
+            'Meta': {'object_name': 'Group'},
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '80'}),
+            'permissions': ('django.db.models.fields.related.ManyToManyField', [], {'to': u"orm['auth.Permission']", 'symmetrical': 'False', 'blank': 'True'})
+        },
+        u'auth.permission': {
+            'Meta': {'ordering': "(u'content_type__app_label', u'content_type__model', u'codename')", 'unique_together': "((u'content_type', u'codename'),)", 'object_name': 'Permission'},
+            'codename': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'content_type': ('django.db.models.fields.related.ForeignKey', [], {'to': u"orm['contenttypes.ContentType']"}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '50'})
+        },
+        u'auth.user': {
+            'Meta': {'object_name': 'User'},
+            'date_joined': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'email': ('django.db.models.fields.EmailField', [], {'max_length': '75', 'blank': 'True'}),
+            'first_name': ('django.db.models.fields.CharField', [], {'max_length': '30', 'blank': 'True'}),
+            'groups': ('django.db.models.fields.related.ManyToManyField', [], {'symmetrical': 'False', 'related_name': "u'user_set'", 'blank': 'True', 'to': u"orm['auth.Group']"}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'is_active': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'is_staff': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'is_superuser': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'last_login': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'last_name': ('django.db.models.fields.CharField', [], {'max_length': '30', 'blank': 'True'}),
+            'password': ('django.db.models.fields.CharField', [], {'max_length': '128'}),
+            'user_permissions': ('django.db.models.fields.related.ManyToManyField', [], {'symmetrical': 'False', 'related_name': "u'user_set'", 'blank': 'True', 'to': u"orm['auth.Permission']"}),
+            'username': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '30'})
+        },
+        u'contenttypes.contenttype': {
+            'Meta': {'ordering': "('name',)", 'unique_together': "(('app_label', 'model'),)", 'object_name': 'ContentType', 'db_table': "'django_content_type'"},
+            'app_label': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'model': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '100'})
+        },
+        u'filer.file': {
+            'Meta': {'object_name': 'File'},
+            '_file_size': ('django.db.models.fields.IntegerField', [], {'null': 'True', 'blank': 'True'}),
+            'description': ('django.db.models.fields.TextField', [], {'null': 'True', 'blank': 'True'}),
+            'file': ('django.db.models.fields.files.FileField', [], {'max_length': '255', 'null': 'True', 'blank': 'True'}),
+            'folder': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "u'all_files'", 'null': 'True', 'to': u"orm['filer.Folder']"}),
+            'has_all_mandatory_data': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'is_public': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'modified_at': ('django.db.models.fields.DateTimeField', [], {'auto_now': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'default': "u''", 'max_length': '255', 'blank': 'True'}),
+            'original_filename': ('django.db.models.fields.CharField', [], {'max_length': '255', 'null': 'True', 'blank': 'True'}),
+            'owner': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "u'owned_files'", 'null': 'True', 'to': u"orm['auth.User']"}),
+            'polymorphic_ctype': ('django.db.models.fields.related.ForeignKey', [], {'related_name': "u'polymorphic_filer.file_set'", 'null': 'True', 'to': u"orm['contenttypes.ContentType']"}),
+            'sha1': ('django.db.models.fields.CharField', [], {'default': "u''", 'max_length': '40', 'blank': 'True'}),
+            'uploaded_at': ('django.db.models.fields.DateTimeField', [], {'auto_now_add': 'True', 'blank': 'True'})
+        },
+        u'filer.folder': {
+            'Meta': {'ordering': "(u'name',)", 'unique_together': "((u'parent', u'name'),)", 'object_name': 'Folder'},
+            'created_at': ('django.db.models.fields.DateTimeField', [], {'auto_now_add': 'True', 'blank': 'True'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            u'level': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'}),
+            u'lft': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'}),
+            'modified_at': ('django.db.models.fields.DateTimeField', [], {'auto_now': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '255'}),
+            'owner': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "u'filer_owned_folders'", 'null': 'True', 'to': u"orm['auth.User']"}),
+            'parent': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "u'children'", 'null': 'True', 'to': u"orm['filer.Folder']"}),
+            u'rght': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'}),
+            u'tree_id': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'}),
+            'uploaded_at': ('django.db.models.fields.DateTimeField', [], {'auto_now_add': 'True', 'blank': 'True'})
+        },
+        'filer.image': {
+            'Meta': {'object_name': 'Image', '_ormbases': [u'filer.File']},
+            '_height': ('django.db.models.fields.IntegerField', [], {'null': 'True', 'blank': 'True'}),
+            '_width': ('django.db.models.fields.IntegerField', [], {'null': 'True', 'blank': 'True'}),
+            'author': ('django.db.models.fields.CharField', [], {'max_length': '255', 'null': 'True', 'blank': 'True'}),
+            'date_taken': ('django.db.models.fields.DateTimeField', [], {'null': 'True', 'blank': 'True'}),
+            'default_alt_text': ('django.db.models.fields.CharField', [], {'max_length': '255', 'null': 'True', 'blank': 'True'}),
+            'default_caption': ('django.db.models.fields.CharField', [], {'max_length': '255', 'null': 'True', 'blank': 'True'}),
+            u'file_ptr': ('django.db.models.fields.related.OneToOneField', [], {'to': u"orm['filer.File']", 'unique': 'True', 'primary_key': 'True'}),
+            'must_always_publish_author_credit': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'must_always_publish_copyright': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'subject_location': ('django.db.models.fields.CharField', [], {'default': 'None', 'max_length': '64', 'null': 'True', 'blank': 'True'})
+        },
+        u'socialaggregator.aggregator': {
+            'Meta': {'object_name': 'Aggregator'},
+            'creation_date': ('django.db.models.fields.DateTimeField', [], {'auto_now_add': 'True', 'blank': 'True'}),
+            'feeds': ('django.db.models.fields.related.ManyToManyField', [], {'to': u"orm['socialaggregator.Feed']", 'symmetrical': 'False'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '250'}),
+            'query': ('django.db.models.fields.CharField', [], {'max_length': '250'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'unique': 'True', 'max_length': '100'}),
+            'social_plugin': ('django.db.models.fields.CharField', [], {'max_length': '250'})
+        },
+        u'socialaggregator.feed': {
+            'Meta': {'object_name': 'Feed'},
+            'creation_date': ('django.db.models.fields.DateTimeField', [], {'auto_now_add': 'True', 'blank': 'True'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '250'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'unique': 'True', 'max_length': '100'})
+        },
+        u'socialaggregator.ressource': {
+            'Meta': {'ordering': "('-priority', 'name')", 'object_name': 'Ressource'},
+            'activate': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'author': ('django.db.models.fields.CharField', [], {'max_length': '250'}),
+            'background_color': ('django.db.models.fields.CharField', [], {'max_length': '250', 'blank': 'True'}),
+            'button_color': ('django.db.models.fields.CharField', [], {'default': "'black'", 'max_length': '100'}),
+            'button_label': ('django.db.models.fields.CharField', [], {'max_length': '100', 'blank': 'True'}),
+            'creation_date': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime(2014, 10, 16, 0, 0)'}),
+            'description': ('django.db.models.fields.TextField', [], {'blank': 'True'}),
+            'favorite': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'feeds': ('django.db.models.fields.related.ManyToManyField', [], {'to': u"orm['socialaggregator.Feed']", 'symmetrical': 'False'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'image': ('django.db.models.fields.related.ForeignKey', [], {'default': 'None', 'related_name': "'edsa_ressource_image'", 'null': 'True', 'blank': 'True', 'to': "orm['filer.Image']"}),
+            'language': ('django.db.models.fields.CharField', [], {'max_length': '2', 'blank': 'True'}),
+            'media_url': ('django.db.models.fields.URLField', [], {'max_length': '500', 'blank': 'True'}),
+            'media_url_type': ('django.db.models.fields.CharField', [], {'max_length': '100', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '250'}),
+            'new_page': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'priority': ('django.db.models.fields.IntegerField', [], {'default': '100'}),
+            'query': ('django.db.models.fields.CharField', [], {'max_length': '250', 'blank': 'True'}),
+            'ressource_date': ('django.db.models.fields.DateTimeField', [], {}),
+            'short_description': ('django.db.models.fields.TextField', [], {'blank': 'True'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'unique': 'True', 'max_length': '100'}),
+            'social_id': ('django.db.models.fields.CharField', [], {'max_length': '250', 'blank': 'True'}),
+            'social_type': ('django.db.models.fields.CharField', [], {'default': "'edsa_article'", 'max_length': '250'}),
+            'text_display': ('django.db.models.fields.CharField', [], {'default': "'default'", 'max_length': '100'}),
+            'thumbnail': ('django.db.models.fields.related.ForeignKey', [], {'default': 'None', 'related_name': "'edsa_ressource_thumbnail'", 'null': 'True', 'blank': 'True', 'to': "orm['filer.Image']"}),
+            'update_date': ('django.db.models.fields.DateTimeField', [], {'default': 'None'}),
+            'updated': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'view_size': ('django.db.models.fields.CharField', [], {'default': "'default'", 'max_length': '100'})
+        }
+    }
+
+    complete_apps = ['socialaggregator']